requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
mysql-connector-python>=8.0.0
tqdm>=4.65.0
requests>=2.28.0
